feat(cli): Modularize CLI package structure (PR2)
Create src/cli/ package and extract configuration commands as the first
modularization step toward reducing cli.py to <100 lines.

Changes:
- Create src/cli/__init__.py - Package entry point with re-exports
- Create src/cli/main.py - Main CLI group definition (cli, setup_logging)
- Create src/cli/utils.py - Shared utilities (format_option, project root, etc.)
- Create src/cli/config.py - Config commands (config group, init-config, hello)
- Create src/cli/__main__.py - Module execution support for python -m src.cli
- Rename src/cli.py to src/cli_main.py (Python prefers packages over modules)
- Update pyproject.toml entry points to use src.cli_main:cli

Technical notes:
- Python prefers packages over modules when both exist, so renamed cli.py
- src/cli/__init__.py imports from cli_main for backward compatibility
- All 794 tests passing, Pylint 10.00/10, Xenon A-grade complexity
- DRY violations in linter commands expected (addressed in PR3)

Part of: Consultant Grade Improvements roadmap (PR2 of 6)
